Add storage of outages
Outage information is stored in `outages.yaml`.

Below is an example outage entry, which specifies that there are an outage for 2 services, from 2024-05-22 12:00:00 to 2024-05-29T03:00:00 Zulu (UTC) time:

- url: https://nerc.mghpcc.org/event/mghpcc-annual-power-shutdown-2024/
  timeframes:
    - from: "2024-05-22T12:00:00Z"
      until: "2024-05-29T03:00:00Z"
      affected_services:
        - NERC OpenStack
        - NERC OpenShift

The following are validation rules for the outages and rates are specified in the updated `README.md`

The pydantic models for outages and rates have been moved to a `models` folder for better organization.

New test cases have been added.